+++ lib/middlewares/src/metering.rs
@@ -8,10 +8,8 @@
 //! [See the `metering` detailed and complete
 //! example](https://github.com/wasmerio/wasmer/blob/master/examples/metering.rs).
 
-use loupe::{MemoryUsage, MemoryUsageTracker};
 use std::convert::TryInto;
 use std::fmt;
-use std::mem;
 use std::sync::{Arc, Mutex};
 use wasmer::wasmparser::{Operator, Type as WpType, TypeOrFuncType as WpTypeOrFuncType};
 use wasmer::{
@@ -20,7 +18,7 @@ use wasmer::{
 };
 use wasmer_types::{GlobalIndex, ModuleInfo};
 
-#[derive(Clone, MemoryUsage)]
+#[derive(Clone)]
 struct MeteringGlobalIndexes(GlobalIndex, GlobalIndex);
 
 impl MeteringGlobalIndexes {
@@ -195,13 +193,6 @@ impl<F: Fn(&Operator) -> u64 + Send + Sync + 'static> ModuleMiddleware for Meter
             remaining_points_global_index,
             points_exhausted_global_index,
         ))
-    }
-}
-
-impl<F: Fn(&Operator) -> u64 + Send + Sync + 'static> MemoryUsage for Metering<F> {
-    fn size_of_val(&self, tracker: &mut dyn MemoryUsageTracker) -> usize {
-        mem::size_of_val(self) + self.global_indexes.size_of_val(tracker)
-            - mem::size_of_val(&self.global_indexes)
     }
 }