+++ lib/vm/src/table.rs
@@ -9,7 +9,6 @@ use crate::func_data_registry::VMFuncRef;
 use crate::trap::{Trap, TrapCode};
 use crate::vmcontext::VMTableDefinition;
 use crate::VMExternRef;
-use loupe::{MemoryUsage, MemoryUsageTracker};
 #[cfg(feature = "enable-rkyv")]
 use rkyv::{Archive, Deserialize as RkyvDeserialize, Serialize as RkyvSerialize};
 use serde::{Deserialize, Serialize};
@@ -22,7 +21,7 @@ use std::sync::Mutex;
 use wasmer_types::{ExternRef, TableType, Type as ValType};
 
 /// Implementation styles for WebAssembly tables.
-#[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize, MemoryUsage)]
+#[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
 #[cfg_attr(
     feature = "enable-rkyv",
     derive(RkyvSerialize, RkyvDeserialize, Archive)
@@ -33,7 +32,7 @@ pub enum TableStyle {
 }
 
 /// Trait for implementing the interface of a Wasm table.
-pub trait Table: fmt::Debug + Send + Sync + MemoryUsage {
+pub trait Table: fmt::Debug + Send + Sync {
     /// Returns the style for this Table.
     fn style(&self) -> &TableStyle;
 
@@ -148,12 +147,6 @@ fn table_element_size_test() {
     assert_eq!(size_of::<RawTableElement>(), size_of::<VMFuncRef>());
 }
 
-impl MemoryUsage for RawTableElement {
-    fn size_of_val(&self, _: &mut dyn MemoryUsageTracker) -> usize {
-        std::mem::size_of_val(self)
-    }
-}
-
 impl fmt::Debug for RawTableElement {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_struct("RawTableElement").finish()
@@ -175,7 +168,7 @@ impl Default for TableElement {
 }
 
 /// A table instance.
-#[derive(Debug, MemoryUsage)]
+#[derive(Debug)]
 pub struct LinearTable {
     // TODO: we can remove the mutex by using atomic swaps and preallocating the max table size
     vec: Mutex<Vec<RawTableElement>>,
@@ -189,7 +182,7 @@ pub struct LinearTable {
 
 /// A type to help manage who is responsible for the backing table of the
 /// `VMTableDefinition`.
-#[derive(Debug, MemoryUsage)]
+#[derive(Debug)]
 enum VMTableDefinitionOwnership {
     /// The `VMTableDefinition` is owned by the `Instance` and we should use
     /// its table. This is how a local table that's exported should be stored.