+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Looking for changes that affect our C API? See the [C API Changelog](lib/c-api/C
 
 ## **Unreleased**
 
+### Changed
+- #2868 Removed loupe crate dependency
+
 ### Fixed
 - [#2829](https://github.com/wasmerio/wasmer/pull/2829) Improve error message oriented from JS object.
 - [#2828](https://github.com/wasmerio/wasmer/pull/2828) Fix JsImportObject resolver.